Move registry one way up
